# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Dukanen Marketplace API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface MessageResponse {
  message: string;
}

export interface RegisterInput {
  /** @minLength 2 */
  name: string;
  email: string;
  /** @minLength 6 */
  password: string;
  phone: string;
  location?: string;
  isSeller?: boolean;
}

export interface LoginInput {
  email: string;
  password: string;
}

export type UserRole = typeof UserRole[keyof typeof UserRole];


export const UserRole = {
  buyer: 'buyer',
  seller: 'seller',
  admin: 'admin',
} as const;

export interface User {
  id: number;
  name: string;
  email: string;
  phone: string;
  /** @nullable */
  location?: string | null;
  /** @nullable */
  avatarUrl?: string | null;
  role: UserRole;
  isBanned: boolean;
  createdAt: string;
}

export interface AuthResponse {
  user: User;
  token: string;
}

export interface UserProfile {
  id: number;
  name: string;
  phone: string;
  /** @nullable */
  location?: string | null;
  /** @nullable */
  avatarUrl?: string | null;
  role: string;
  createdAt: string;
  productCount?: number;
}

export interface UserUpdate {
  name?: string;
  phone?: string;
  location?: string;
  avatarUrl?: string;
}

export interface Category {
  id: number;
  name: string;
  slug: string;
  icon: string;
  productCount?: number;
}

export type ProductCondition = typeof ProductCondition[keyof typeof ProductCondition];


export const ProductCondition = {
  new: 'new',
  used: 'used',
  refurbished: 'refurbished',
} as const;

export type ProductStatus = typeof ProductStatus[keyof typeof ProductStatus];


export const ProductStatus = {
  active: 'active',
  sold: 'sold',
  removed: 'removed',
  pending: 'pending',
} as const;

export interface Product {
  id: number;
  title: string;
  description: string;
  price: number;
  currency: string;
  condition: ProductCondition;
  status: ProductStatus;
  /** @nullable */
  location?: string | null;
  imageUrls?: string[];
  categoryId: number;
  /** @nullable */
  categoryName?: string | null;
  sellerId: number;
  /** @nullable */
  sellerName?: string | null;
  /** @nullable */
  sellerPhone?: string | null;
  viewCount?: number;
  createdAt: string;
}

export type ProductInputCondition = typeof ProductInputCondition[keyof typeof ProductInputCondition];


export const ProductInputCondition = {
  new: 'new',
  used: 'used',
  refurbished: 'refurbished',
} as const;

export interface ProductInput {
  /** @minLength 3 */
  title: string;
  /** @minLength 10 */
  description: string;
  /** @minimum 0 */
  price: number;
  currency: string;
  condition: ProductInputCondition;
  location?: string;
  imageUrls?: string[];
  categoryId: number;
}

export type ProductUpdateCondition = typeof ProductUpdateCondition[keyof typeof ProductUpdateCondition];


export const ProductUpdateCondition = {
  new: 'new',
  used: 'used',
  refurbished: 'refurbished',
} as const;

export type ProductUpdateStatus = typeof ProductUpdateStatus[keyof typeof ProductUpdateStatus];


export const ProductUpdateStatus = {
  active: 'active',
  sold: 'sold',
  removed: 'removed',
  pending: 'pending',
} as const;

export interface ProductUpdate {
  title?: string;
  description?: string;
  price?: number;
  currency?: string;
  condition?: ProductUpdateCondition;
  location?: string;
  imageUrls?: string[];
  categoryId?: number;
  status?: ProductUpdateStatus;
}

export interface ProductListResponse {
  products: Product[];
  total: number;
  page: number;
  limit: number;
}

export interface ActivityItem {
  type: string;
  description: string;
  createdAt: string;
}

export interface SellerDashboard {
  totalListings: number;
  activeListings: number;
  soldListings: number;
  totalViews: number;
  unreadMessages: number;
  recentActivity: ActivityItem[];
}

export interface Conversation {
  id: number;
  productId: number;
  /** @nullable */
  productTitle?: string | null;
  /** @nullable */
  productImageUrl?: string | null;
  buyerId: number;
  /** @nullable */
  buyerName?: string | null;
  sellerId: number;
  /** @nullable */
  sellerName?: string | null;
  /** @nullable */
  lastMessage?: string | null;
  /** @nullable */
  lastMessageAt?: string | null;
  unreadCount?: number;
  createdAt: string;
}

export interface ConversationInput {
  productId: number;
  sellerId: number;
  /** @minLength 1 */
  initialMessage: string;
}

export interface Message {
  id: number;
  conversationId: number;
  senderId: number;
  /** @nullable */
  senderName?: string | null;
  content: string;
  isRead?: boolean;
  createdAt: string;
}

export interface MessageInput {
  /** @minLength 1 */
  content: string;
}

export type ReportInputTargetType = typeof ReportInputTargetType[keyof typeof ReportInputTargetType];


export const ReportInputTargetType = {
  product: 'product',
  user: 'user',
} as const;

export interface ReportInput {
  /** @minLength 5 */
  reason: string;
  targetType: ReportInputTargetType;
  targetId: number;
}

export type ReportStatus = typeof ReportStatus[keyof typeof ReportStatus];


export const ReportStatus = {
  pending: 'pending',
  resolved: 'resolved',
  dismissed: 'dismissed',
} as const;

export interface Report {
  id: number;
  reason: string;
  targetType: string;
  targetId: number;
  reporterId: number;
  status: ReportStatus;
  createdAt: string;
}

export interface AdminStats {
  totalUsers: number;
  totalSellers: number;
  totalProducts: number;
  activeProducts: number;
  totalReports: number;
  pendingReports: number;
  newUsersThisWeek: number;
  newProductsThisWeek: number;
}

export interface AdminUserListResponse {
  users: User[];
  total: number;
  page: number;
  limit: number;
}

export type AdminUserUpdateRole = typeof AdminUserUpdateRole[keyof typeof AdminUserUpdateRole];


export const AdminUserUpdateRole = {
  buyer: 'buyer',
  seller: 'seller',
  admin: 'admin',
} as const;

export interface AdminUserUpdate {
  isBanned?: boolean;
  role?: AdminUserUpdateRole;
}

export interface AdminProductListResponse {
  products: Product[];
  total: number;
  page: number;
  limit: number;
}

export type AdminProductUpdateStatus = typeof AdminProductUpdateStatus[keyof typeof AdminProductUpdateStatus];


export const AdminProductUpdateStatus = {
  active: 'active',
  sold: 'sold',
  removed: 'removed',
  pending: 'pending',
} as const;

export interface AdminProductUpdate {
  status?: AdminProductUpdateStatus;
}

export interface AdminReportListResponse {
  reports: Report[];
  total: number;
  page: number;
  limit: number;
}

export type AdminReportUpdateStatus = typeof AdminReportUpdateStatus[keyof typeof AdminReportUpdateStatus];


export const AdminReportUpdateStatus = {
  pending: 'pending',
  resolved: 'resolved',
  dismissed: 'dismissed',
} as const;

export interface AdminReportUpdate {
  status?: AdminReportUpdateStatus;
}

export type ListProductsParams = {
search?: string;
category?: string;
location?: string;
minPrice?: number;
maxPrice?: number;
page?: number;
limit?: number;
};

export type ListAdminUsersParams = {
page?: number;
limit?: number;
search?: string;
};

export type ListAdminProductsParams = {
page?: number;
limit?: number;
status?: string;
};

export type ListAdminReportsParams = {
page?: number;
limit?: number;
status?: string;
};