# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Dukanen Marketplace API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /products:
    get:
      operationId: listProducts
      tags: [products]
      summary: List/search products
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: category
          in: query
          schema:
            type: string
        - name: location
          in: query
          schema:
            type: string
        - name: minPrice
          in: query
          schema:
            type: number
        - name: maxPrice
          in: query
          schema:
            type: number
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductListResponse"
    post:
      operationId: createProduct
      tags: [products]
      summary: Create a product listing
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /products/{id}:
    get:
      operationId: getProduct
      tags: [products]
      summary: Get product by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateProduct
      tags: [products]
      summary: Update a product listing
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
    delete:
      operationId: deleteProduct
      tags: [products]
      summary: Delete a product listing
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"

  # ---- USERS ----
  /users/{id}:
    get:
      operationId: getUser
      tags: [users]
      summary: Get public user profile
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateUser
      tags: [users]
      summary: Update user profile
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /users/{id}/products:
    get:
      operationId: getUserProducts
      tags: [users]
      summary: Get products listed by a user
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"

  # ---- SELLER DASHBOARD ----
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
    MessageResponse:
      type: object
      required: [message]
      properties:
        message:
          type: string

    # AUTH
...
    User:
      type: object
      required: [id, name, email, phone, role, isBanned, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        phone:
          type: string
        location:
          type: ["string", "null"]
        avatarUrl:
          type: ["string", "null"]
        role:
          type: string
          enum: [buyer, seller, admin]
        isBanned:
          type: boolean
        createdAt:
          type: string
          format: date-time
    UserProfile:
      type: object
      required: [id, name, phone, role, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        phone:
          type: string
        location:
          type: ["string", "null"]
        avatarUrl:
          type: ["string", "null"]
        role:
          type: string
        createdAt:
          type: string
          format: date-time
        productCount:
          type: integer
    UserUpdate:
      type: object
      properties:
        name:
          type: string
        phone:
          type: string
        location:
          type: string
        avatarUrl:
          type: string

    # CATEGORY
...
    Product:
      type: object
      required: [id, title, description, price, currency, condition, status, categoryId, sellerId, createdAt]
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
        price:
          type: number
        currency:
          type: string
        condition:
          type: string
          enum: [new, used, refurbished]
        status:
          type: string
          enum: [active, sold, removed, pending]
        location:
          type: ["string", "null"]
        imageUrls:
          type: array
          items:
            type: string
        categoryId:
          type: integer
        categoryName:
          type: ["string", "null"]
        sellerId:
          type: integer
        sellerName:
          type: ["string", "null"]
        sellerPhone:
          type: ["string", "null"]
        viewCount:
          type: integer
        createdAt:
          type: string
          format: date-time
    ProductInput:
      type: object
      required: [title, description, price, currency, condition, categoryId]
      properties:
        title:
          type: string
          minLength: 3
        description:
          type: string
          minLength: 10
        price:
          type: number
          minimum: 0
        currency:
          type: string
          default: KES
        condition:
          type: string
          enum: [new, used, refurbished]
        location:
          type: string
        imageUrls:
          type: array
          items:
            type: string
        categoryId:
          type: integer
    ProductUpdate:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        price:
          type: number
        currency:
          type: string
        condition:
          type: string
          enum: [new, used, refurbished]
        location:
          type: string
        imageUrls:
          type: array
          items:
            type: string
        categoryId:
          type: integer
        status:
          type: string
          enum: [active, sold, removed, pending]
    ProductListResponse:
      type: object
      required: [products, total, page, limit]
      properties:
        products:
          type: array
          items:
            $ref: "#/components/schemas/Product"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer

    # SELLER DASHBOARD
...