# lib/api-zod/src/generated/types/productInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Dukanen Marketplace API
 * OpenAPI spec version: 0.1.0
 */
import type { ProductInputCondition } from './productInputCondition';

export interface ProductInput {
  /** @minLength 3 */
  title: string;
  /** @minLength 10 */
  description: string;
  /** @minimum 0 */
  price: number;
  currency: string;
  condition: ProductInputCondition;
  location?: string;
  imageUrls?: string[];
  categoryId: number;
}